src/application/use-cases/educational-content-cases/trail/read/read-not-published/read-not-published-trails-use-case.ts
import {
    ITrailRepository,
    Trail,
    ReadNotPublishedTrailsUseCaseOutputDTO
} from "./index";


export class ReadNotPublishedTrailsUseCase {
    constructor(private trailRepository: ITrailRepository) { }

    async execute(): Promise<ReadNotPublishedTrailsUseCaseOutputDTO> {
        const notPublishedTrails: Trail[] = await this.trailRepository.listNotPublished();

        const outputDTO: ReadNotPublishedTrailsUseCaseOutputDTO = {
            trails: notPublishedTrails
        }

        return outputDTO;
    }
}
